+++ src/org/joml/Matrix3f.java
@@ -562,8 +562,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
         x *= invLength;
         y *= invLength;
         z *= invLength;
-        float s = (float) Math.sin(angle);
-        float c = (float) Math.cosFromSin(s, angle);
+        float s = Math.sin(angle);
+        float c = Math.cosFromSin(s, angle);
         float omc = 1.0f - c;
         m00 = c + x*x*omc;
         m11 = c + y*y*omc;
@@ -1516,8 +1516,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @return this
      */
     public Matrix3f rotation(float angle, float x, float y, float z) {
-        float sin = (float) Math.sin(angle);
-        float cos = (float) Math.cosFromSin(sin, angle);
+        float sin = Math.sin(angle);
+        float cos = Math.cosFromSin(sin, angle);
         float C = 1.0f - cos;
         float xy = x * y, xz = x * z, yz = y * z;
         m00 = cos + x * x * C;
@@ -1547,8 +1547,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      */
     public Matrix3f rotationX(float ang) {
         float sin, cos;
-        sin = (float) Math.sin(ang);
-        cos = (float) Math.cosFromSin(sin, ang);
+        sin = Math.sin(ang);
+        cos = Math.cosFromSin(sin, ang);
         m00 = 1.0f;
         m01 = 0.0f;
         m02 = 0.0f;
@@ -1576,8 +1576,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      */
     public Matrix3f rotationY(float ang) {
         float sin, cos;
-        sin = (float) Math.sin(ang);
-        cos = (float) Math.cosFromSin(sin, ang);
+        sin = Math.sin(ang);
+        cos = Math.cosFromSin(sin, ang);
         m00 = cos;
         m01 = 0.0f;
         m02 = -sin;
@@ -1605,8 +1605,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      */
     public Matrix3f rotationZ(float ang) {
         float sin, cos;
-        sin = (float) Math.sin(ang);
-        cos = (float) Math.cosFromSin(sin, ang);
+        sin = Math.sin(ang);
+        cos = Math.cosFromSin(sin, ang);
         m00 = cos;
         m01 = sin;
         m02 = 0.0f;
@@ -1638,12 +1638,12 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @return this
      */
     public Matrix3f rotationXYZ(float angleX, float angleY, float angleZ) {
-        float sinX = (float) Math.sin(angleX);
-        float cosX = (float) Math.cosFromSin(sinX, angleX);
-        float sinY = (float) Math.sin(angleY);
-        float cosY = (float) Math.cosFromSin(sinY, angleY);
-        float sinZ = (float) Math.sin(angleZ);
-        float cosZ = (float) Math.cosFromSin(sinZ, angleZ);
+        float sinX = Math.sin(angleX);
+        float cosX = Math.cosFromSin(sinX, angleX);
+        float sinY = Math.sin(angleY);
+        float cosY = Math.cosFromSin(sinY, angleY);
+        float sinZ = Math.sin(angleZ);
+        float cosZ = Math.cosFromSin(sinZ, angleZ);
         float m_sinX = -sinX;
         float m_sinY = -sinY;
         float m_sinZ = -sinZ;
@@ -1689,12 +1689,12 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @return this
      */
     public Matrix3f rotationZYX(float angleZ, float angleY, float angleX) {
-        float sinX = (float) Math.sin(angleX);
-        float cosX = (float) Math.cosFromSin(sinX, angleX);
-        float sinY = (float) Math.sin(angleY);
-        float cosY = (float) Math.cosFromSin(sinY, angleY);
-        float sinZ = (float) Math.sin(angleZ);
-        float cosZ = (float) Math.cosFromSin(sinZ, angleZ);
+        float sinX = Math.sin(angleX);
+        float cosX = Math.cosFromSin(sinX, angleX);
+        float sinY = Math.sin(angleY);
+        float cosY = Math.cosFromSin(sinY, angleY);
+        float sinZ = Math.sin(angleZ);
+        float cosZ = Math.cosFromSin(sinZ, angleZ);
         float m_sinZ = -sinZ;
         float m_sinY = -sinY;
         float m_sinX = -sinX;
@@ -1740,12 +1740,12 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @return this
      */
     public Matrix3f rotationYXZ(float angleY, float angleX, float angleZ) {
-        float sinX = (float) Math.sin(angleX);
-        float cosX = (float) Math.cosFromSin(sinX, angleX);
-        float sinY = (float) Math.sin(angleY);
-        float cosY = (float) Math.cosFromSin(sinY, angleY);
-        float sinZ = (float) Math.sin(angleZ);
-        float cosZ = (float) Math.cosFromSin(sinZ, angleZ);
+        float sinX = Math.sin(angleX);
+        float cosX = Math.cosFromSin(sinX, angleX);
+        float sinY = Math.sin(angleY);
+        float cosY = Math.cosFromSin(sinY, angleY);
+        float sinZ = Math.sin(angleZ);
+        float cosZ = Math.cosFromSin(sinZ, angleZ);
         float m_sinY = -sinY;
         float m_sinX = -sinX;
         float m_sinZ = -sinZ;
@@ -1895,8 +1895,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      */
     public Matrix3f rotateX(float ang, Matrix3f dest) {
         float sin, cos;
-        sin = (float) Math.sin(ang);
-        cos = (float) Math.cosFromSin(sin, ang);
+        sin = Math.sin(ang);
+        cos = Math.cosFromSin(sin, ang);
         float rm11 = cos;
         float rm21 = -sin;
         float rm12 = sin;
@@ -1948,8 +1948,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      */
     public Matrix3f rotateY(float ang, Matrix3f dest) {
         float sin, cos;
-        sin = (float) Math.sin(ang);
-        cos = (float) Math.cosFromSin(sin, ang);
+        sin = Math.sin(ang);
+        cos = Math.cosFromSin(sin, ang);
         float rm00 = cos;
         float rm20 = sin;
         float rm02 = -sin;
@@ -2001,8 +2001,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      */
     public Matrix3f rotateZ(float ang, Matrix3f dest) {
         float sin, cos;
-        sin = (float) Math.sin(ang);
-        cos = (float) Math.cosFromSin(sin, ang);
+        sin = Math.sin(ang);
+        cos = Math.cosFromSin(sin, ang);
         float rm00 = cos;
         float rm10 = -sin;
         float rm01 = sin;
@@ -2103,12 +2103,12 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @see org.joml.Matrix3fc#rotateXYZ(float, float, float, org.joml.Matrix3f)
      */
     public Matrix3f rotateXYZ(float angleX, float angleY, float angleZ, Matrix3f dest) {
-        float sinX = (float) Math.sin(angleX);
-        float cosX = (float) Math.cosFromSin(sinX, angleX);
-        float sinY = (float) Math.sin(angleY);
-        float cosY = (float) Math.cosFromSin(sinY, angleY);
-        float sinZ = (float) Math.sin(angleZ);
-        float cosZ = (float) Math.cosFromSin(sinZ, angleZ);
+        float sinX = Math.sin(angleX);
+        float cosX = Math.cosFromSin(sinX, angleX);
+        float sinY = Math.sin(angleY);
+        float cosY = Math.cosFromSin(sinY, angleY);
+        float sinZ = Math.sin(angleZ);
+        float cosZ = Math.cosFromSin(sinZ, angleZ);
         float m_sinX = -sinX;
         float m_sinY = -sinY;
         float m_sinZ = -sinZ;
@@ -2191,12 +2191,12 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @see org.joml.Matrix3fc#rotateZYX(float, float, float, org.joml.Matrix3f)
      */
     public Matrix3f rotateZYX(float angleZ, float angleY, float angleX, Matrix3f dest) {
-        float sinX = (float) Math.sin(angleX);
-        float cosX = (float) Math.cosFromSin(sinX, angleX);
-        float sinY = (float) Math.sin(angleY);
-        float cosY = (float) Math.cosFromSin(sinY, angleY);
-        float sinZ = (float) Math.sin(angleZ);
-        float cosZ = (float) Math.cosFromSin(sinZ, angleZ);
+        float sinX = Math.sin(angleX);
+        float cosX = Math.cosFromSin(sinX, angleX);
+        float sinY = Math.sin(angleY);
+        float cosY = Math.cosFromSin(sinY, angleY);
+        float sinZ = Math.sin(angleZ);
+        float cosZ = Math.cosFromSin(sinZ, angleZ);
         float m_sinZ = -sinZ;
         float m_sinY = -sinY;
         float m_sinX = -sinX;
@@ -2279,12 +2279,12 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @see org.joml.Matrix3fc#rotateYXZ(float, float, float, org.joml.Matrix3f)
      */
     public Matrix3f rotateYXZ(float angleY, float angleX, float angleZ, Matrix3f dest) {
-        float sinX = (float) Math.sin(angleX);
-        float cosX = (float) Math.cosFromSin(sinX, angleX);
-        float sinY = (float) Math.sin(angleY);
-        float cosY = (float) Math.cosFromSin(sinY, angleY);
-        float sinZ = (float) Math.sin(angleZ);
-        float cosZ = (float) Math.cosFromSin(sinZ, angleZ);
+        float sinX = Math.sin(angleX);
+        float cosX = Math.cosFromSin(sinX, angleX);
+        float sinY = Math.sin(angleY);
+        float cosY = Math.cosFromSin(sinY, angleY);
+        float sinZ = Math.sin(angleZ);
+        float cosZ = Math.cosFromSin(sinZ, angleZ);
         float m_sinY = -sinY;
         float m_sinX = -sinX;
         float m_sinZ = -sinZ;
@@ -2348,8 +2348,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @see org.joml.Matrix3fc#rotate(float, float, float, float, org.joml.Matrix3f)
      */
     public Matrix3f rotate(float ang, float x, float y, float z, Matrix3f dest) {
-        float s = (float) Math.sin(ang);
-        float c = (float) Math.cosFromSin(s, ang);
+        float s = Math.sin(ang);
+        float c = Math.cosFromSin(s, ang);
         float C = 1.0f - c;
 
         // rotation matrix elements:
@@ -2424,8 +2424,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @return dest
      */
     public Matrix3f rotateLocal(float ang, float x, float y, float z, Matrix3f dest) {
-        float s = (float) Math.sin(ang);
-        float c = (float) Math.cosFromSin(s, ang);
+        float s = Math.sin(ang);
+        float c = Math.cosFromSin(s, ang);
         float C = 1.0f - c;
         float xx = x * x, xy = x * y, xz = x * z;
         float yy = y * y, yz = y * z;
@@ -2523,8 +2523,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @return dest
      */
     public Matrix3f rotateLocalX(float ang, Matrix3f dest) {
-        float sin = (float) Math.sin(ang);
-        float cos = (float) Math.cosFromSin(sin, ang);
+        float sin = Math.sin(ang);
+        float cos = Math.cosFromSin(sin, ang);
         float nm01 = cos * m01 - sin * m02;
         float nm02 = sin * m01 + cos * m02;
         float nm11 = cos * m11 - sin * m12;
@@ -2597,8 +2597,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @return dest
      */
     public Matrix3f rotateLocalY(float ang, Matrix3f dest) {
-        float sin = (float) Math.sin(ang);
-        float cos = (float) Math.cosFromSin(sin, ang);
+        float sin = Math.sin(ang);
+        float cos = Math.cosFromSin(sin, ang);
         float nm00 =  cos * m00 + sin * m02;
         float nm02 = -sin * m00 + cos * m02;
         float nm10 =  cos * m10 + sin * m12;
@@ -2671,8 +2671,8 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @return dest
      */
     public Matrix3f rotateLocalZ(float ang, Matrix3f dest) {
-        float sin = (float) Math.sin(ang);
-        float cos = (float) Math.cosFromSin(sin, ang);
+        float sin = Math.sin(ang);
+        float cos = Math.cosFromSin(sin, ang);
         float nm00 = cos * m00 - sin * m01;
         float nm01 = sin * m00 + cos * m01;
         float nm10 = cos * m10 - sin * m11;
@@ -3750,9 +3750,9 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @see org.joml.Matrix3fc#getScale(org.joml.Vector3f)
      */
     public Vector3f getScale(Vector3f dest) {
-        dest.x = (float) Math.sqrt(m00 * m00 + m01 * m01 + m02 * m02);
-        dest.y = (float) Math.sqrt(m10 * m10 + m11 * m11 + m12 * m12);
-        dest.z = (float) Math.sqrt(m20 * m20 + m21 * m21 + m22 * m22);
+        dest.x = Math.sqrt(m00 * m00 + m01 * m01 + m02 * m02);
+        dest.y = Math.sqrt(m10 * m10 + m11 * m11 + m12 * m12);
+        dest.z = Math.sqrt(m20 * m20 + m21 * m21 + m22 * m22);
         return dest;
     }
 
@@ -4318,9 +4318,9 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @return dest
      */
     public Vector3f getEulerAnglesZYX(Vector3f dest) {
-        dest.x = (float) Math.atan2(m12, m22);
-        dest.y = (float) Math.atan2(-m02, (float) Math.sqrt(m12 * m12 + m22 * m22));
-        dest.z = (float) Math.atan2(m01, m00);
+        dest.x = Math.atan2(m12, m22);
+        dest.y = Math.atan2(-m02, Math.sqrt(m12 * m12 + m22 * m22));
+        dest.z = Math.atan2(m01, m00);
         return dest;
     }
 
@@ -4497,12 +4497,12 @@ public class Matrix3f implements Externalizable, Matrix3fc {
      * @see org.joml.Matrix3fc#reflect(org.joml.Quaternionfc, org.joml.Matrix3f)
      */
     public Matrix3f reflect(Quaternionfc orientation, Matrix3f dest) {
-        double num1 = orientation.x() + orientation.x();
-        double num2 = orientation.y() + orientation.y();
-        double num3 = orientation.z() + orientation.z();
-        float normalX = (float) (orientation.x() * num3 + orientation.w() * num2);
-        float normalY = (float) (orientation.y() * num3 - orientation.w() * num1);
-        float normalZ = (float) (1.0 - (orientation.x() * num1 + orientation.y() * num2));
+        float num1 = orientation.x() + orientation.x();
+        float num2 = orientation.y() + orientation.y();
+        float num3 = orientation.z() + orientation.z();
+        float normalX = orientation.x() * num3 + orientation.w() * num2;
+        float normalY = orientation.y() * num3 - orientation.w() * num1;
+        float normalZ = 1.0f - (orientation.x() * num1 + orientation.y() * num2);
         return reflect(normalX, normalY, normalZ, dest);
     }
